minor feedback addressed
